pyproject.toml
[tool.poetry]
name = "agentops-studio"
version = "0.1.0"
description = "Production-grade system for designing, running, and observing LLM agents"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
gunicorn = "^21.2.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"
langchain-anthropic = "^0.1.1"
langchain-google-genai = "^0.0.6"
langsmith = "^0.0.77"
openai = "^1.10.0"
anthropic = "^0.9.0"
google-generativeai = "^0.3.2"
redis = {extras = ["hiredis"], version = "^5.0.1"}
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
boto3 = "^1.34.27"
sentry-sdk = {extras = ["fastapi"], version = "^1.40.0"}
python-multipart = "^0.0.6"
prometheus-client = "^0.19.0"
structlog = "^24.1.0"
tenacity = "^8.2.3"
phonenumbers = "^8.13.27"
presidio-analyzer = "^2.2.354"
presidio-anonymizer = "^2.2.354"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
ruff = "^0.1.11"
mypy = "^1.8.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
